+++ packages/react-ui/src/index.d.ts
@@ -34,6 +34,11 @@ import AppBar from './components/organisms/AppBar/AppBar';
 import LabelWithIndicator from './components/atoms/LabelWithIndicator';
 import { getCartoColorStylePropsForItem } from './utils/palette';
 import Avatar from './components/molecules/Avatar';
+import {
+  ICON_SIZE_SMALL,
+  ICON_SIZE_MEDIUM,
+  ICON_SIZE_LARGE
+} from './theme/themeconstants';
 
 export {
   theme,
@@ -72,5 +77,8 @@ export {
   AppBar,
   LabelWithIndicator,
   getCartoColorStylePropsForItem,
-  Avatar
+  Avatar,
+  ICON_SIZE_SMALL,
+  ICON_SIZE_MEDIUM,
+  ICON_SIZE_LARGE
 };