+++ packages/react-ui/src/theme/sections/components/dataDisplay.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ICON_SIZE, ICON_SIZE_M } from '../../themeConstants';
+import { ICON_SIZE_MEDIUM, ICON_SIZE_LARGE, ICON_SIZE_SMALL } from '../../themeConstants';
 import { getSpacing } from '../../themeUtils';
 import { commonPalette } from '../palette';
 import { themeTypography } from '../typography';
@@ -75,7 +75,7 @@ export const dataDisplayOverrides = {
         marginRight: getSpacing(1),
 
         '& .MuiSvgIcon-root': {
-          fontSize: ICON_SIZE_M
+          fontSize: ICON_SIZE_LARGE
         },
         '.MuiMenuItem-root.MuiButtonBase-root &': {
           minWidth: getSpacing(2.25)
@@ -91,7 +91,7 @@ export const dataDisplayOverrides = {
           width: getSpacing(4)
         },
         '& .MuiSvgIcon-root': {
-          fontSize: ICON_SIZE
+          fontSize: ICON_SIZE_MEDIUM
         }
       }
     }
@@ -272,8 +272,8 @@ export const dataDisplayOverrides = {
         padding: getSpacing(0, 0.5),
 
         '& .MuiAvatar-root': {
-          width: ICON_SIZE_M,
-          height: ICON_SIZE_M,
+          width: ICON_SIZE_LARGE,
+          height: ICON_SIZE_LARGE,
           margin: 0,
           color: commonPalette.secondary.contrastText,
           backgroundColor: commonPalette.background.paper,
@@ -284,8 +284,8 @@ export const dataDisplayOverrides = {
           marginLeft: getSpacing(0.5)
         },
         '& img': {
-          width: ICON_SIZE_M,
-          height: ICON_SIZE_M
+          width: ICON_SIZE_LARGE,
+          height: ICON_SIZE_LARGE
         },
         '&.Mui-disabled': {
           color: commonPalette.text.disabled,
@@ -344,12 +344,12 @@ export const dataDisplayOverrides = {
       // Sizes
       sizeSmall: {
         '& img': {
-          width: ICON_SIZE,
-          height: ICON_SIZE
+          width: ICON_SIZE_MEDIUM,
+          height: ICON_SIZE_MEDIUM
         },
         '& .MuiAvatar-root': {
-          width: ICON_SIZE,
-          height: ICON_SIZE
+          width: ICON_SIZE_MEDIUM,
+          height: ICON_SIZE_MEDIUM
         },
         '& .MuiChip-icon': {
           marginLeft: getSpacing(0.25)
@@ -367,8 +367,8 @@ export const dataDisplayOverrides = {
         padding: getSpacing(0, 0.5)
       },
       deleteIcon: {
-        width: ICON_SIZE,
-        height: ICON_SIZE,
+        width: ICON_SIZE_MEDIUM,
+        height: ICON_SIZE_MEDIUM,
         margin: 0,
         marginLeft: '2px', // Forced to a non-standard value to meet with design
         marginRight: '3px', // Forced to a non-standard value to meet with design
@@ -483,10 +483,13 @@ export const dataDisplayOverrides = {
   MuiSvgIcon: {
     styleOverrides: {
       root: {
-        fontSize: ICON_SIZE
+        fontSize: ICON_SIZE_MEDIUM
       },
-      sizeLarge: {
-        fontSize: ICON_SIZE_M
+      fontSizeSmall: {
+        fontSize: ICON_SIZE_SMALL
+      },
+      fontSizeLarge: {
+        fontSize: ICON_SIZE_LARGE
       }
     }
   }