+++ packages/react-ui/storybook/.storybook/preview.js
@@ -104,6 +104,8 @@ export const parameters = {
         'Introduction',
         'Foundations',
         ['Palette', 'Typography', 'Spacing', 'Borders', 'Elevations', 'Breakpoints'],
+        'Icons',
+        ['Introduction', 'Carto Icons'],
         'Atoms',
         'Molecules',
         'Organisms'